edição variaveis

+++ tipos-variaveis/src/TiposVariaveis.java
@@ -1,7 +1,7 @@
 public class TiposVariaveis {
     public static void main(String[] args) {
         short numeroCurto = 1;
-        short numeroNormal = numeroCurto;
+        int numeroNormal = numeroCurto;
         short numeroCurto2 = (short) numeroNormal;
 
     }
